refactor(ports): expand core/ports/stt.py with TranscriptionResult, errors, helpers
Move TranscriptionResult, STTUnavailableError, STTNoiseError, _WHISPER_NOISE_TOKENS,
is_whisper_noise, mime_from_suffix inline into core/ports/stt.py. Remove
TYPE_CHECKING cross-reference to lyra.stt. Update lyra.stt re-export shim to
source all types from core/ports/stt (eliminates duplicate-class typecheck errors).
Dead code (STTConfig, load_stt_config) not included in core/ports.
